add: documentation, packaging
