+++ src/cyclic_operations/__init__.py
@@ -1,0 +1,15 @@
+from src.cyclic_operations.C4 import C4_Cyclic_Permutation, C4_Pool, C4_Roll, C4_Slice, C4_Stack
+from src.cyclic_operations.D4 import D4_Cyclic_Permutation, D4_Pool, D4_Roll, D4_Slice, D4_Stack
+
+__all__ = [
+    "C4_Cyclic_Permutation",
+    "C4_Pool",
+    "C4_Roll",
+    "C4_Slice",
+    "C4_Stack",
+    "D4_Cyclic_Permutation",
+    "D4_Pool",
+    "D4_Roll",
+    "D4_Slice",
+    "D4_Stack",
+]
